backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
#pydantic==2.4.2 // in fastapi
#transformers==4.35.0 // on unsloth
#torch==2.1.0 #unsloth==2025.5.6 requires torch>=2.4.0
#torch==2.4.0
torch==2.4.0
#accelerate==0.24.0 #unsloth 2025.5.6 depends on accelerate>=0.34.1
accelerate==0.34.1
bitsandbytes==0.41.1
# unsloth @ git+https://github.com/unslothai/unsloth.git?
unsloth==2025.5.6
python-dotenv==1.0.0
#sentencepiece==0.1.99 #unsloth==2025.5.6 requires sentencepiece>=0.2.0
#sentencepiece==0.2.0
sentencepiece==0.2.0
#protobuf==4.25.1 #unsloth version 2025.5.6 requires an older version of protobuf
protobuf==3.20.3
einops==0.7.0
#safetensors==0.4.1 // in accelerate
runpod==1.3.0 